+++ src/Cryptography/Shares.cs
@@ -90,7 +90,7 @@ namespace SecretSharingDotNet.Cryptography
         /// Gets the original secret
         /// </summary>
         /// <remarks>Legacy property</remarks>
-        [Obsolete("Legacy property. Will be removed in futures versions.", false)]
+        [Obsolete("Legacy property. Will be removed in futures versions. Pleas use OriginalSecret property.", true)]
         public Secret<TNumber> Item1 => this.OriginalSecret;
 
         /// <summary>
@@ -105,7 +105,7 @@ namespace SecretSharingDotNet.Cryptography
         /// Gets the shares.
         /// </summary>
         /// <remarks>Legacy property</remarks>
-        [Obsolete("Legacy property. Will be removed in futures versions.", false)]
+        [Obsolete("Legacy property. Will be removed in futures versions.", true)]
         public ICollection<FinitePoint<TNumber>> Item2 => this.shareList;
 
         /// <summary>
@@ -118,6 +118,7 @@ namespace SecretSharingDotNet.Cryptography
         /// </summary>
         /// <param name="shares">A <see cref="Shares{TNumber}"/> object.</param>
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Design", "CA1006:DoNotNestGenericTypesInMemberSignatures")]
+        [Obsolete("Legacy property. Will be removed in futures versions.", true)]
         public static implicit operator Tuple<Secret<TNumber>, ICollection<FinitePoint<TNumber>>>(Shares<TNumber> shares) => new Tuple<Secret<TNumber>, ICollection<FinitePoint<TNumber>>>(shares?.OriginalSecret, shares);
 
         /// <summary>
@@ -225,7 +226,7 @@ namespace SecretSharingDotNet.Cryptography
         {
             if (this.IsReadOnly)
             {
-                throw new NotSupportedException($"The {nameof(Shares<TNumber>)} collection is read-only.");
+                throw new NotSupportedException(string.Format(ErrorMessages.ReadOnlyCollection, nameof(Shares<TNumber>)));
             }
 
             this.shareList.Clear();
@@ -242,7 +243,7 @@ namespace SecretSharingDotNet.Cryptography
         {
             if (this.IsReadOnly)
             {
-                throw new NotSupportedException($"The {nameof(Shares<TNumber>)} collection is read-only.");
+                throw new NotSupportedException(string.Format(ErrorMessages.ReadOnlyCollection, nameof(Shares<TNumber>)));
             }
 
             if (!this.Contains(item))
@@ -263,7 +264,7 @@ namespace SecretSharingDotNet.Cryptography
         {
             if (this.IsReadOnly)
             {
-                throw new NotSupportedException($"The {nameof(Shares<TNumber>)} collection is read-only.");
+                throw new NotSupportedException(string.Format(ErrorMessages.ReadOnlyCollection, nameof(Shares<TNumber>)));
             }
 
             return this.shareList.Remove(item);
@@ -284,7 +285,7 @@ namespace SecretSharingDotNet.Cryptography
                     this.CopyTo(x, arrayIndex);
                     break;
                 default:
-                    throw new InvalidCastException($"Parameter {nameof(array)}: Cannot convert an array from type {array.GetType().GetElementType()} to an array of type {typeof(FinitePoint<TNumber>)}.");
+                    throw new InvalidCastException(string.Format(ErrorMessages.InvalidArrayTypeCast, nameof(array), array.GetType().GetElementType(), typeof(FinitePoint<TNumber>)));
             }
         }
 
@@ -301,12 +302,12 @@ namespace SecretSharingDotNet.Cryptography
             _ = array ?? throw new ArgumentNullException(nameof(array));
             if (arrayIndex < 0)
             {
-                throw new ArgumentOutOfRangeException(nameof(arrayIndex), "The starting array index cannot be negative.");
+                throw new ArgumentOutOfRangeException(nameof(arrayIndex), ErrorMessages.StartArrayIndexNegative);
             }
 
             if (Count > array.Length - arrayIndex + 1)
             {
-                throw new ArgumentException("The destination array has fewer elements than the collection.", nameof(array));
+                throw new ArgumentException(ErrorMessages.DestinationArrayHasFewerElements, nameof(array));
             }
 
             for (int i = 0; i < this.shareList.Count; i++)