+++ tests/ShamirsSecretSharingTest.cs
@@ -67,27 +67,6 @@ namespace SecretSharingDotNet.Test
         }
 
         /// <summary>
-        /// Tests the legacy usage of shares
-        /// </summary>
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Performance", "CA1822:MarkMembersAsStatic")]
-        [Fact]
-        public void LegacyShareUsage()
-        {
-            var split = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>());
-            var combine = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>());
-            var shares = split.MakeShares(3, 7, TestData.DefaultTestPassword);
-            var secret = shares.Item1;
-            var subSet1 = shares.Item2.Where(p => p.X.IsEven).ToList();
-            var recoveredSecret1 = combine.Reconstruction(subSet1.ToArray());
-            var subSet2 = shares.Item2.Where(p => !p.X.IsEven).ToList();
-            var recoveredSecret2 = combine.Reconstruction(subSet2.ToArray());
-            Assert.Equal(TestData.DefaultTestPassword, recoveredSecret1);
-            Assert.Equal(secret, recoveredSecret1);
-            Assert.Equal(secret, recoveredSecret2);
-            Assert.Equal(521, split.SecurityLevel);
-        }
-
-        /// <summary>
         /// Tests the security level auto-detection of <see cref="ShamirsSecretSharing{TNumber}"/>.
         /// </summary>
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Performance", "CA1822:MarkMembersAsStatic")]
@@ -145,10 +124,9 @@ namespace SecretSharingDotNet.Test
         [MemberData(nameof(TestData.TestPasswordData), MemberType = typeof(TestData))]
         public void TestWithPassword(int splitSecurityLevel, int expectedSecurityLevel, string password)
         {
-            var split = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>(),
-                splitSecurityLevel);
+            var split = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>());
             var combine = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>());
-            var shares = split.MakeShares(3, 7, password);
+            var shares = split.MakeShares(3, 7, password, splitSecurityLevel);
             Assert.True(shares.OriginalSecretExists);
             Assert.NotNull(shares.OriginalSecret);
             var secret = shares.OriginalSecret;
@@ -173,10 +151,9 @@ namespace SecretSharingDotNet.Test
         [MemberData(nameof(TestData.TestNumberData), MemberType = typeof(TestData))]
         public void TestWithNumber(int splitSecurityLevel, int expectedSecurityLevel, BigInteger number)
         {
-            var split = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>(),
-                splitSecurityLevel);
+            var split = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>());
             var combine = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>());
-            var shares = split.MakeShares(3, 7, number);
+            var shares = split.MakeShares(3, 7, number, splitSecurityLevel);
             Assert.True(shares.OriginalSecretExists);
             Assert.NotNull(shares.OriginalSecret);
             var secret = shares.OriginalSecret;
@@ -194,19 +171,15 @@ namespace SecretSharingDotNet.Test
         /// Tests <see cref="ShamirsSecretSharing{TNumber}"/> with random <see cref="BigInteger"/> value as secret.
         /// </summary>
         /// <param name="splitSecurityLevel">Initial security level for secret split phase</param>
-        /// <param name="combineSecurityLevel">Initial security level for secret combine phase (aka reconstruction phase)</param>
         /// <param name="expectedSecurityLevel">Expected security level after secret reconstruction</param>
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Performance", "CA1822:MarkMembersAsStatic")]
         [Theory]
         [MemberData(nameof(TestData.TestRandomSecretData), MemberType = typeof(TestData))]
-        public void TestWithRandomSecret(int splitSecurityLevel, int combineSecurityLevel, int expectedSecurityLevel)
+        public void TestWithRandomSecret(int splitSecurityLevel, int expectedSecurityLevel)
         {
-            var split = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>(),
-                splitSecurityLevel);
-            var combine =
-                new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>(),
-                    combineSecurityLevel);
-            var shares = split.MakeShares(3, 7);
+            var split = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>());
+            var combine = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>());
+            var shares = split.MakeShares(3, 7, splitSecurityLevel);
             Assert.True(shares.OriginalSecretExists);
             Assert.NotNull(shares.OriginalSecret);
             var secret = shares.OriginalSecret;
@@ -227,8 +200,8 @@ namespace SecretSharingDotNet.Test
         [Fact]
         public void TestMinimumSharedSecretsMake()
         {
-            var sss = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>(), 5);
-            Assert.Throws<ArgumentOutOfRangeException>(() => sss.MakeShares(1, 7));
+            var sss = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>());
+            Assert.Throws<ArgumentOutOfRangeException>(() => sss.MakeShares(1, 7, 5));
         }
 
         /// <summary>
@@ -238,8 +211,8 @@ namespace SecretSharingDotNet.Test
         [Fact]
         public void TestMinimumSharedSecretsReconstruction()
         {
-            var sss = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>(), 13);
-            var shares = sss.MakeShares(2, 7);
+            var sss = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>());
+            var shares = sss.MakeShares(2, 7, 13);
             var subSet = shares.Where(p => p.X == Calculator<BigInteger>.One).ToList();
             Assert.Throws<ArgumentOutOfRangeException>(() => sss.Reconstruction(subSet.ToArray()));
         }
@@ -251,23 +224,11 @@ namespace SecretSharingDotNet.Test
         [Fact]
         public void TestShareThreshold()
         {
-            var sss = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>(), 51);
-            var shares = sss.MakeShares(3, 7);
+            var sss = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>());
+            var shares = sss.MakeShares(3, 7, 51);
             var subSet = shares.Take(2).ToList();
             var secret = sss.Reconstruction(subSet.ToArray());
             Assert.NotEqual(shares.OriginalSecret, secret);
-        }
-
-        /// <summary>
-        /// Tests whether or not the <see cref="InvalidOperationException"/> is thrown if the security level
-        /// is not initialized.
-        /// </summary>
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Performance", "CA1822:MarkMembersAsStatic")]
-        [Fact]
-        public void TestUninitializedSecurityLevel()
-        {
-            var sss = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>());
-            Assert.Throws<InvalidOperationException>(() => sss.MakeShares(2, 7));
         }
 
         /// <summary>
@@ -279,9 +240,9 @@ namespace SecretSharingDotNet.Test
         {
             const string longSecret =
                 "-----BEGIN EC PRIVATE KEY-----MIIBUQIBAQQgxq7AWG9L6uleuTB9q5FGqnHjXF+kD4y9154SLYYKMDqggeMwgeACAQEwLAYHKoZIzj0BAQIhAP////////////////////////////////////7///wvMEQEIAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABCAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABwRBBHm+Zn753LusVaBilc6HCwcCm/zbLc4o2VnygVsW+BeYSDradyajxGVdpPv8DhEIqP0XtEimhVQZnEfQj/sQ1LgCIQD////////////////////+uq7c5q9IoDu/0l6M0DZBQQIBAaFEA0IABE0XO6I8lZYzXqRQnHP/knSwLex7q77g4J2AN0cVyrADicGlUr6QjVIlIu9NXCHxD2i++ToWjO1zLVdxgNJbUUc=-----END EC PRIVATE KEY-----";
-            var split = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>(), 1024);
-            var combine = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>(), 5);
-            var shares = split.MakeShares(3, 7, longSecret);
+            var split = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>());
+            var combine = new ShamirsSecretSharing<BigInteger>(new ExtendedEuclideanAlgorithm<BigInteger>());
+            var shares = split.MakeShares(3, 7, longSecret, 1024);
             var subSet1 = shares.Where(p => p.X.IsEven).ToList();
             var recoveredSecret1 = combine.Reconstruction(subSet1.ToArray());
             var subSet2 = shares.Where(p => !p.X.IsEven).ToList();